// src/app/data/interfaces/form-interfaces.ts
export interface RegistrationFormValues {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  birthDate: string;
  country: string;
  city: string;
  postalCode: string;
  address: string;
  shippingAddress: string;
  isDefaultAddress: boolean;
  billingAddress: string;
  isSameAddress: boolean;
}